pgsql close connection fix
PHP @ used as an error control operator (to silence error echo output) is
deprecated. By not suppressing the error output in the disconnect method (which
it does to return a custom raiseError()) errors are echoed in an unexpected way
now in PHP 8.

Change approach to try/catch instead of $ok = @pg_closed.